refactor: remove duplicate registrar helper
